requirements.txt
asgiref==3.8.1
colorama==0.4.6
coloredlogs==15.0.1
contourpy==1.3.2
cycler==0.12.1
Django==4.2
django-filter==23.2
djangorestframework==3.14.0
filelock==3.18.0
flatbuffers==25.2.10
fonttools==4.58.2
fsspec==2025.5.1
humanfriendly==10.0
imageio==2.37.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
lime==0.2.0.1
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.5
numpy==2.1.0
onnxruntime==1.22.0
packaging==25.0
pillow==11.2.1
protobuf==6.31.0
psycopg2-binary==2.9.10
pyparsing==3.2.3
pyreadline3==3.5.4
python-dateutil==2.9.0.post0
pytz==2025.2
scikit-image==0.25.2
scikit-learn==1.7.0
scipy==1.15.3
setuptools==80.8.0
six==1.17.0
sqlparse==0.5.3
sympy==1.14.0
threadpoolctl==3.6.0
tifffile==2025.6.1
torch==2.7.1
torchvision==0.22.1
tqdm==4.67.1
typing_extensions==4.14.0
tzdata==2025.2
reportlab==4.2.5
python-docx==1.1.2
